@@ -9,11 +9,11 @@
   "scripts": {
     "clean": "rimraf lib build",
     "lint": "eslint . --ext .ts --quiet",
-    "build": "tsc -p tsconfig.json",
-    "watch": "tsc -p tsconfig.json --watch true --sourceMap",
-    "test": "node --trace-warnings node_modules/jest/bin/jest.js --runInBand --detectOpenHandles --forceExit",
+    "build": "webpack",
+    "watch": "webpack -w",
+    "test": "node --trace-warnings node_modules/jest/bin/jest.js --runInBand --forceExit",
     "test-build": "node --trace-warnings node_modules/jest/bin/jest.js --runInBand --coverage --forceExit",
-    "prepare": "npm-run-all clean build"
+    "prepare": "webpack"
   },
   "repository": {
     "type": "git",
@@ -33,6 +33,9 @@
     "globals": {
       "__TEST__": true
     },
+    "projects": [
+      "<rootDir>"
+    ],
     "watchman": false,
     "clearMocks": true,
     "globalSetup": "./jest.js",
@@ -54,36 +57,35 @@
     "@types/debounce": "^3.0.0",
     "@types/fb-watchman": "^2.0.0",
     "@types/glob": "^7.1.3",
-    "@types/jest": "^26.0.15",
-    "@types/marked": "^1.1.0",
+    "@types/jest": "^26.0.18",
+    "@types/marked": "^1.2.1",
     "@types/minimatch": "^3.0.3",
     "@types/mkdirp": "^1.0.1",
     "@types/node": "^10.12.0",
     "@types/semver": "^7.3.4",
-    "@types/tar": "^4.0.3",
+    "@types/tar": "^4.0.4",
     "@types/uuid": "^8.3.0",
     "@types/which": "^1.3.2",
-    "@typescript-eslint/eslint-plugin": "^4.6.0",
-    "@typescript-eslint/eslint-plugin-tslint": "^4.6.0",
-    "@typescript-eslint/parser": "^4.6.0",
+    "@typescript-eslint/eslint-plugin": "^4.9.1",
+    "@typescript-eslint/parser": "^4.9.1",
     "colors": "^1.4.0",
-    "eslint": "^7.12.1",
-    "eslint-config-prettier": "^6.15.0",
-    "eslint-plugin-jest": "^24.1.0",
-    "eslint-plugin-jsdoc": "^30.7.3",
-    "jest": "26.6.1",
-    "npm-run-all": "^4.1.5",
-    "prettier": "^2.1.2",
-    "ts-jest": "^26.4.3",
-    "typescript": "^4.0.5",
-    "vscode-languageserver": "^6.1.1"
+    "eslint": "^7.15.0",
+    "eslint-plugin-jest": "^24.1.3",
+    "eslint-plugin-jsdoc": "^30.7.8",
+    "jest": "26.6.3",
+    "ts-jest": "^26.4.4",
+    "ts-loader": "^8.0.11",
+    "typescript": "^4.1.2",
+    "vscode-languageserver": "next",
+    "webpack": "^5.11.0",
+    "webpack-cli": "^4.2.0"
   },
   "dependencies": {
-    "@chemzqm/neovim": "^5.2.9",
-    "ansi-styles": "^4.3.0",
+    "@chemzqm/neovim": "^5.2.12",
+    "ansi-styles": "^5.0.0",
     "bser": "^2.1.1",
     "bytes": "^3.1.0",
-    "cli-table": "^0.3.1",
+    "cli-table": "^0.3.4",
     "clipboardy": "^2.3.0",
     "content-disposition": "^0.5.3",
     "debounce": "^1.2.0",
@@ -97,7 +99,7 @@
     "isuri": "^2.0.3",
     "jsonc-parser": "^2.3.1",
     "log4js": "^6.3.0",
-    "marked": "^1.2.3",
+    "marked": "^1.2.5",
     "minimatch": "^3.0.4",
     "promise.prototype.finally": "^3.1.2",
     "rc": "^1.2.8",
@@ -106,6 +108,7 @@
     "tslib": "^2.0.3",
     "unzipper": "^0.10.11",
     "uuid": "^7.0.3",
+    "vscode-jsonrpc": "^5.0.1",
     "vscode-languageserver-protocol": "^3.15.3",
     "vscode-languageserver-textdocument": "^1.0.1",
     "vscode-languageserver-types": "^3.15.1",
