@@ -1,6 +1,7 @@
-const semver = require('semver')
-const version = process.version.replace('v', '')
+import semver from 'semver'
 const promiseFinally = require('promise.prototype.finally')
+
+const version = process.version.replace('v', '')
 if (!semver.gte(version, '8.10.0')) {
   console.error('node version ' + version + ' < 8.10.0, please upgrade nodejs, or use `let g:coc_node_path = "/path/to/node"` in your vimrc')
   process.exit()
@@ -11,23 +12,23 @@ if (!semver.gte(version, '10.12.0')) {
   }
 }
 Object.defineProperty(console, 'log', {
-  value: function () {
+  value: function() {
     logger.info(...arguments)
   }
 })
 promiseFinally.shim()
-const logger = require('../lib/util/logger')('server')
-const attach = require('../lib/attach').default
+const logger = require('./util/logger')('server')
+const attach = require('./attach').default
 
-attach({reader: process.stdin, writer: process.stdout})
+attach({ reader: process.stdin, writer: process.stdout })
 
-process.on('uncaughtException', function (err) {
+process.on('uncaughtException', function(err) {
   let msg = 'Uncaught exception: ' + err.message
   console.error(msg)
   logger.error('uncaughtException', err.stack)
 })
 
-process.on('unhandledRejection', function (reason, p) {
+process.on('unhandledRejection', function(reason, p) {
   if (reason instanceof Error) {
     console.error('UnhandledRejection: ' + reason.message + '\n' + reason.stack)
   } else {
