@@ -3,7 +3,7 @@ import { Neovim } from '@chemzqm/neovim'
 import { Color, Range, CancellationToken, ColorInformation, ColorPresentation, Disposable } from 'vscode-languageserver-protocol'
 import { TextDocument } from 'vscode-languageserver-textdocument'
 import Colors from '../../handler/colors'
-import { toHexString } from '../../handler/highlighter'
+import { toHexString } from '../../handler/helper'
 import languages from '../../languages'
 import { ProviderResult } from '../../provider'
 import { disposeAll } from '../../util'
