@@ -11,8 +11,8 @@ import { TextDocumentContentProvider } from '../../provider'
 import { ConfigurationTarget } from '../../types'
 import { disposeAll } from '../../util'
 import { readFile } from '../../util/fs'
-import workspace from '../../workspace'
 import window from '../../window'
+import workspace from '../../workspace'
 import helper, { createTmpFile } from '../helper'
 
 let nvim: Neovim
@@ -884,7 +884,7 @@ describe('workspace events', () => {
     let doc = await helper.createDocument()
     let filepath = URI.parse(doc.uri).fsPath
     let fn = jest.fn()
-    let disposable = workspace.onWillSaveUntil(event => {
+    let disposable = workspace.onWillSaveTextDocument(event => {
       let promise = new Promise<TextEdit[]>(resolve => {
         fn()
         let edit: TextEdit = {
@@ -894,7 +894,7 @@ describe('workspace events', () => {
         resolve([edit])
       })
       event.waitUntil(promise)
-    }, null, 'test')
+    })
     await helper.wait(100)
     await nvim.setLine('bar')
     await helper.wait(30)
@@ -904,6 +904,71 @@ describe('workspace events', () => {
     expect(content.startsWith('foobar')).toBe(true)
     disposable.dispose()
     expect(fn).toBeCalledTimes(1)
+    if (fs.existsSync(filepath)) {
+      fs.unlinkSync(filepath)
+    }
+  })
+
+  it('should not work for async waitUntil', async () => {
+    let doc = await helper.createDocument()
+    let filepath = URI.parse(doc.uri).fsPath
+    let disposable = workspace.onWillSaveTextDocument(event => {
+      setTimeout(() => {
+        let edit: TextEdit = {
+          newText: 'foo',
+          range: Range.create(0, 0, 0, 0)
+        }
+        event.waitUntil(Promise.resolve([edit]))
+      }, 30)
+    })
+    await nvim.setLine('bar')
+    await helper.wait(30)
+    await nvim.command('wa')
+    let content = doc.getDocumentContent()
+    expect(content).toMatch('bar')
+    disposable.dispose()
+    if (fs.existsSync(filepath)) {
+      fs.unlinkSync(filepath)
+    }
+  })
+
+  it('should only use first returned textEdits', async () => {
+    let doc = await helper.createDocument()
+    let filepath = URI.parse(doc.uri).fsPath
+    let disposables: Disposable[] = []
+    workspace.onWillSaveTextDocument(event => {
+      event.waitUntil(Promise.resolve(undefined))
+    }, null, disposables)
+    workspace.onWillSaveTextDocument(event => {
+      let promise = new Promise<TextEdit[]>(resolve => {
+        setTimeout(() => {
+          let edit: TextEdit = {
+            newText: 'foo',
+            range: Range.create(0, 0, 0, 0)
+          }
+          resolve([edit])
+        }, 10)
+      })
+      event.waitUntil(promise)
+    }, null, disposables)
+    workspace.onWillSaveTextDocument(event => {
+      let promise = new Promise<TextEdit[]>(resolve => {
+        setTimeout(() => {
+          let edit: TextEdit = {
+            newText: 'bar',
+            range: Range.create(0, 0, 0, 0)
+          }
+          resolve([edit])
+        }, 30)
+      })
+      event.waitUntil(promise)
+    }, null, disposables)
+    await nvim.setLine('bar')
+    await helper.wait(30)
+    await nvim.command('wa')
+    let content = doc.getDocumentContent()
+    expect(content).toMatch('foo')
+    disposeAll(disposables)
     if (fs.existsSync(filepath)) {
       fs.unlinkSync(filepath)
     }
@@ -962,5 +1027,33 @@ describe('workspace textDocument content provider', () => {
     let buf = await nvim.buffer
     let lines = await buf.lines
     expect(lines).toEqual(['bar'])
+  })
+})
+
+describe('workspace registerBufferSync', () => {
+  it('should regist', async () => {
+    await helper.createDocument()
+    let created = 0
+    let deleted = 0
+    let changed = 0
+    let disposable = workspace.registerBufferSync(() => {
+      created = created + 1
+      return {
+        dispose: () => {
+          deleted += 1
+        },
+        onChange: () => {
+          changed += 1
+        }
+      }
+    })
+    disposables.push(disposable)
+    let doc = await helper.createDocument()
+    expect(created).toBe(2)
+    await doc.applyEdits([TextEdit.insert(Position.create(0, 0), 'foo')])
+    expect(changed).toBe(1)
+    await nvim.command('bd!')
+    await helper.wait(50)
+    expect(deleted).toBe(1)
   })
 })
