@@ -1,13 +1,12 @@
 import { Buffer, Neovim } from '@chemzqm/neovim'
 import debounce from 'debounce'
 import { CancellationToken, CancellationTokenSource, Color, ColorInformation, Disposable, Position, Range } from 'vscode-languageserver-protocol'
-import Document from '../model/document'
-import { wait } from '../util'
+import languages from '../languages'
 import { group } from '../util/array'
 import { equals } from '../util/object'
 import { positionInRange } from '../util/position'
 import workspace from '../workspace'
-import languages from '../languages'
+import { isDark, toHexColor, toHexString } from './helper'
 const logger = require('../util/logger')('highlighter')
 
 export interface ColorRanges {
@@ -47,16 +46,13 @@ export default class Highlighter implements Disposable {
     let doc = workspace.getDocument(this.bufnr)
     if (!doc) return
     try {
-      this.cancel()
       this.tokenSource = new CancellationTokenSource()
       let { token } = this.tokenSource
-      await synchronizeDocument(doc)
       if (workspace.insertMode) return
       if (token.isCancellationRequested) return
       if (this.version && doc.version == this.version) return
       let { version } = doc
       let colors: ColorInformation[]
-      usedColors.clear()
       colors = await languages.provideDocumentColors(doc.textDocument, token)
       colors = colors || []
       if (token.isCancellationRequested) return
@@ -155,43 +151,5 @@ export default class Highlighter implements Disposable {
       this.tokenSource.dispose()
       this.tokenSource = null
     }
-  }
-}
-
-export function toHexString(color: Color): string {
-  let c = toHexColor(color)
-  return `${pad(c.red.toString(16))}${pad(c.green.toString(16))}${pad(c.blue.toString(16))}`
-}
-
-function pad(str: string): string {
-  return str.length == 1 ? `0${str}` : str
-}
-
-function toHexColor(color: Color): { red: number; green: number; blue: number } {
-  let { red, green, blue } = color
-  return {
-    red: Math.round(red * 255),
-    green: Math.round(green * 255),
-    blue: Math.round(blue * 255)
-  }
-}
-
-function isDark(color: Color): boolean {
-  // http://www.w3.org/TR/WCAG20/#relativeluminancedef
-  let rgb = [color.red, color.green, color.blue]
-  let lum = []
-  for (let i = 0; i < rgb.length; i++) {
-    let chan = rgb[i]
-    lum[i] = (chan <= 0.03928) ? chan / 12.92 : Math.pow(((chan + 0.055) / 1.055), 2.4)
-  }
-  let luma = 0.2126 * lum[0] + 0.7152 * lum[1] + 0.0722 * lum[2]
-  return luma <= 0.5
-}
-
-async function synchronizeDocument(doc: Document): Promise<void> {
-  let { changedtick } = doc
-  await doc.patchChange()
-  if (changedtick != doc.changedtick) {
-    await wait(50)
   }
 }
