@@ -19,14 +19,35 @@ import BasicList from './list/basic'
 import diagnosticManager from './diagnostic/manager'
 import { ansiparse } from './util/ansiparse'
 import Watchman from './watchman'
+import { Mutex } from './util/mutex'
 import { URI } from 'vscode-uri'
 import { Neovim, Buffer, Window } from '@chemzqm/neovim'
-import { Disposable, Event, Emitter } from 'vscode-languageserver-protocol'
+import {
+  Disposable,
+  Position,
+  Range,
+  TextEdit,
+  RequestType,
+  RequestType0,
+  NotificationType,
+  NotificationType0,
+  Event,
+  CancellationToken,
+  CancellationTokenSource,
+  Emitter,
+  Diagnostic,
+  DiagnosticSeverity,
+  CompletionItemKind,
+  InsertTextFormat,
+  Location,
+  LocationLink,
+} from 'vscode-languageserver-protocol'
+import { ProgressType } from 'vscode-jsonrpc'
 
 export * from './types'
 export * from './language-client'
 export * from './provider'
 
-export { Neovim, Buffer, Window, Highligher, Mru, Watchman, URI as Uri, Disposable, Event, Emitter, FloatFactory, fetch, download, ansiparse }
-export { workspace, window, snippetManager, events, services, commands, sources, languages, diagnosticManager, Document, FileSystemWatcher, extensions, listManager, BasicList }
-export { disposeAll, concurrent, wait, runCommand, isRunning, executable } from './util'
+export { Neovim, Diagnostic, DiagnosticSeverity, CompletionItemKind, InsertTextFormat, Location, LocationLink, CancellationTokenSource, CancellationToken, ProgressType, Position, Range, TextEdit, RequestType, RequestType0, NotificationType, NotificationType0, Buffer, Window, Highligher, Mru, Watchman, URI as Uri, Disposable, Event, Emitter, FloatFactory, fetch, download, ansiparse }
+export { workspace, window, snippetManager, events, services, commands, sources, languages, diagnosticManager, Document, FileSystemWatcher, extensions, listManager, BasicList, Mutex }
+export { disposeAll, concurrent, watchFile, wait, runCommand, isRunning, executable } from './util'
