@@ -250,6 +250,10 @@ export class LanguageClient extends BaseLanguageClient {
 
   private checkProcessDied(childProcess: ChildProcess | undefined): void {
     if (!childProcess || global.hasOwnProperty('__TEST__')) return
+    if (global.hasOwnProperty('__TEST__')) {
+      process.kill(childProcess.pid, 0)
+      return
+    }
     setTimeout(() => {
       // Test if the process is still alive. Throws an exception if not
       try {
@@ -258,7 +262,7 @@ export class LanguageClient extends BaseLanguageClient {
       } catch (error) {
         // All is fine.
       }
-    }, 1000)
+    }, 2000)
   }
 
   protected handleConnectionClosed(): void {
@@ -383,6 +387,9 @@ export class LanguageClient extends BaseLanguageClient {
             return Promise.reject<MessageTransports>(`Launching server module "${node.module}" failed.`)
           }
           logger.info(`${this.id} started with ${serverProcess.pid}`)
+          serverProcess.on('error', e => {
+            logger.error(`Process ${runtime} error: `, e)
+          })
           this._serverProcess = serverProcess
           serverProcess.stderr.on('data', data => this.appendOutput(data, encoding))
           return {
@@ -483,10 +490,6 @@ export class LanguageClient extends BaseLanguageClient {
 
   private appendOutput(data: any, encoding: string): void {
     let msg: string = Is.string(data) ? data : data.toString(encoding)
-    if (global.hasOwnProperty('__TEST__')) {
-      console.log(msg)
-      return
-    }
     this.outputChannel.append(msg.endsWith('\n') ? msg : msg + '\n')
   }
 }
