@@ -6,26 +6,24 @@
 
 import { Disposable, NotificationHandler, NotificationType, ProgressToken, ProgressType, WorkDoneProgress, WorkDoneProgressBegin, WorkDoneProgressReport } from 'vscode-languageserver-protocol'
 import { StatusBarItem } from '../types'
-import window from '../window'
 import { disposeAll } from '../util'
+import window from '../window'
+const logger = require('../util/logger')('language-client-progressPart')
 
 export interface ProgressContext {
   onProgress<P>(type: ProgressType<P>, token: string | number, handler: NotificationHandler<P>): Disposable
   sendNotification<P, RO>(type: NotificationType<P, RO>, params?: P): void
 }
 
-const progressParts: Map<ProgressToken, ProgressPart> = new Map()
-
-class ProgressPart {
-  private _disposables: Disposable[] = []
-  private _statusBarItem: StatusBarItem | undefined
+export class ProgressPart {
+  private disposables: Disposable[] = []
+  private statusBarItem: StatusBarItem | undefined
   private _cancelled = false
   private title: string
 
-  public constructor(private _client: ProgressContext, private _token: ProgressToken) {
-    this._statusBarItem = window.createStatusBarItem(99, { progress: true })
-    this._disposables.push(this._statusBarItem)
-    this._disposables.push(_client.onProgress(WorkDoneProgress.type, this._token, value => {
+  public constructor(private client: ProgressContext, private token: ProgressToken, done?: (part: ProgressPart) => void) {
+    this.statusBarItem = window.createStatusBarItem(99, { progress: true })
+    this.disposables.push(client.onProgress(WorkDoneProgress.type, this.token, value => {
       switch (value.kind) {
         case 'begin':
           this.begin(value)
@@ -35,22 +33,24 @@ class ProgressPart {
           break
         case 'end':
           this.done(value.message)
+          done && done(this)
           break
       }
     }))
   }
 
   public begin(params: WorkDoneProgressBegin): void {
-    // TODO: support progress window with cancel button & WorkDoneProgressCancelNotification
+    if (typeof this.title === 'string') return
+    // TODO support  params.cancellable
     this.title = params.title
     this.report(params)
   }
 
   private report(params: WorkDoneProgressReport | WorkDoneProgressBegin): void {
-    let statusBarItem = this._statusBarItem
+    let statusBarItem = this.statusBarItem
     let parts: string[] = []
     if (this.title) parts.push(this.title)
-    if (params.percentage) parts.push(params.percentage.toFixed(0) + '%')
+    if (typeof params.percentage == 'number') parts.push(params.percentage.toFixed(0) + '%')
     if (params.message) parts.push(params.message)
     statusBarItem.text = parts.join(' ')
     statusBarItem.show()
@@ -59,42 +59,16 @@ class ProgressPart {
   public cancel(): void {
     if (this._cancelled) return
     this._cancelled = true
-    disposeAll(this._disposables)
-    if (progressParts.has(this._token)) {
-      progressParts.delete(this._token)
-    }
+    disposeAll(this.disposables)
   }
 
   public done(message?: string): void {
-    let statusBarItem = this._statusBarItem
-    if (!message) {
-      this.cancel()
-    } else {
-      statusBarItem.text = `${this.title} ${message}`
-      setTimeout(() => {
-        this.cancel()
-      }, 500)
-    }
+    if (this._cancelled) return
+    const statusBarItem = this.statusBarItem
+    statusBarItem.text = `${this.title} ${message || 'finished'}`
+    setTimeout(() => {
+      statusBarItem.dispose()
+    }, 300)
+    this.cancel()
   }
 }
-
-class ProgressManager {
-  public create(client: ProgressContext, token: ProgressToken): ProgressPart {
-    let part = this.getProgress(token)
-    if (part) return part
-    part = new ProgressPart(client, token)
-    progressParts.set(token, part)
-    return part
-  }
-
-  public getProgress(token: ProgressToken): ProgressPart | null {
-    return progressParts.get(token) || null
-  }
-
-  public cancel(token: ProgressToken): void {
-    let progress = this.getProgress(token)
-    if (progress) progress.cancel()
-  }
-}
-
-export default new ProgressManager()
