@@ -6,10 +6,11 @@ import { URI } from 'vscode-uri'
 import events from '../events'
 import { ChangeInfo, DidChangeTextDocumentParams, Env } from '../types'
 import { diffLines, getChange } from '../util/diff'
-import { isGitIgnored } from '../util/fs'
-import { disposeAll, getUri } from '../util/index'
+import { disposeAll, getUri, wait } from '../util/index'
+import { equals } from '../util/object'
 import { byteLength, byteSlice } from '../util/string'
 import { Chars } from './chars'
+import { Mutex } from '../util/mutex'
 const logger = require('../util/logger')('model-document')
 
 export type LastChangeType = 'insert' | 'change' | 'delete'
@@ -29,21 +30,22 @@ export interface BufferOption {
   lines: string[]
 }
 
-// wrapper class of TextDocument
+// getText, positionAt, offsetAt
 export default class Document {
   public buftype: string
   public isIgnored = false
   public chars: Chars
-  public textDocument: TextDocument
   public fireContentChanges: Function & { clear(): void }
   public fetchContent: Function & { clear(): void }
-  private ignoreChange = false
+  private mutex = new Mutex()
+  private _version = 1
   private size = 0
   private nvim: Neovim
   private eol = true
   private variables: { [key: string]: any }
   // real current lines
-  private lines: string[] = []
+  private lines: ReadonlyArray<string> = []
+  private syncLines: ReadonlyArray<string> = []
   private _attached = false
   private _previewwindow = false
   private _winid = -1
@@ -56,10 +58,7 @@ export default class Document {
   private disposables: Disposable[] = []
   public readonly onDocumentChange: Event<DidChangeTextDocumentParams> = this._onDocumentChange.event
   public readonly onDocumentDetach: Event<number> = this._onDocumentDetach.event
-  constructor(
-    public readonly buffer: Buffer,
-    private env: Env,
-    private maxFileSize: number | null) {
+  constructor(public readonly buffer: Buffer, private env: Env, private maxFileSize: number | null) {
     this.fireContentChanges = debounce(() => {
       this._fireContentChanges()
     }, 100)
@@ -68,6 +67,31 @@ export default class Document {
     }, 100)
   }
 
+  /**
+   * Synchronize content
+   */
+  public get content(): string {
+    return this.syncLines.join('\n') + (this.eol ? '\n' : '')
+  }
+
+  public get version(): number {
+    return this._version
+  }
+
+  /**
+   * Buffer number
+   */
+  public get bufnr(): number {
+    return this.buffer.id
+  }
+
+  public get filetype(): string {
+    return this._filetype
+  }
+
+  public get uri(): string {
+    return this._uri
+  }
   /**
    * Check if current document should be attached for changes.
    *
@@ -131,6 +155,8 @@ export default class Document {
 
   /**
    * Window ID when buffer create, could be -1 when no window associated.
+   *
+   * @deprecated could be wrong.
    */
   public get winid(): number {
     return this._winid
@@ -159,18 +185,17 @@ export default class Document {
     this.variables = opts.variables || {}
     this._changedtick = opts.changedtick
     this.eol = opts.eol == 1
-    let uri = this._uri = getUri(opts.fullpath, this.bufnr, buftype, this.env.isCygwin)
+    this._uri = getUri(opts.fullpath, this.bufnr, buftype, this.env.isCygwin)
     if (token.isCancellationRequested) return false
     if (this.shouldAttach) {
       this.lines = opts.lines
+      this.syncLines = this.lines
       let res = await this.attach()
       if (!res) return false
       this._attached = true
     }
     this._filetype = this.convertFiletype(opts.filetype)
-    this.textDocument = TextDocument.create(uri, this.filetype, 1, this.getDocumentContent())
     this.setIskeyword(opts.iskeyword)
-    this.gitCheck()
     if (token.isCancellationRequested) {
       this.detach()
       return false
@@ -181,43 +206,22 @@ export default class Document {
   private async attach(): Promise<boolean> {
     let attached = await this.buffer.attach(true)
     if (!attached) return false
-    let lastChange: number
-    this.buffer.listen('lines', (...args: any[]) => {
-      // avoid neovim send same change multiple times after checktime
-      if (lastChange == args[1]) return
-      lastChange = args[1]
-      this.onChange.apply(this, args)
-    }, this.disposables)
+    this.buffer.listen('lines', this.onChange.bind(this), this.disposables)
     this.buffer.listen('detach', async buf => {
       this._onDocumentDetach.fire(buf.id)
     }, this.disposables)
-    this.buffer.listen('changedtick', (_buf: Buffer, tick: number) => {
-      this._changedtick = tick
-    }, this.disposables)
-    if (this.textDocument) {
-      this.fireContentChanges()
-    }
     return true
   }
 
-  private onChange(
-    buf: Buffer,
-    tick: number,
-    firstline: number,
-    lastline: number,
-    linedata: string[]
-    // more:boolean
-  ): void {
-    if (buf.id !== this.buffer.id || tick == null) return
-    this._changedtick = tick
-    if (this.ignoreChange) {
-      this.ignoreChange = false
-      return
+  private async onChange(buf: Buffer, tick: number, firstline: number, lastline: number, linedata: string[]): Promise<void> {
+    if (buf.id !== this.bufnr || !this._attached || tick == null) return
+    let release = await this.mutex.acquire()
+    if (tick > this._changedtick) {
+      this._changedtick = tick
+      this.lines = [...this.lines.slice(0, firstline), ...linedata, ...this.lines.slice(lastline)]
+      this.fireContentChanges()
     }
-    let lines = this.lines.slice(0, firstline)
-    lines = lines.concat(linedata, this.lines.slice(lastline))
-    this.lines = lines
-    this.fireContentChanges()
+    release()
   }
 
   /**
@@ -225,35 +229,38 @@ export default class Document {
    */
   public async checkDocument(): Promise<void> {
     let { buffer } = this
+    this.fireContentChanges.clear()
     this._changedtick = await buffer.changedtick
     this.lines = await buffer.lines
-    this.fireContentChanges.clear()
-    this._fireContentChanges()
+    let changed = this._fireContentChanges()
+    if (changed) await wait(30)
   }
 
   /**
    * Check if document changed after last synchronize
    */
   public get dirty(): boolean {
-    return this.content != this.getDocumentContent()
+    if (this.lines === this.syncLines) return false
+    return !equals(this.lines, this.syncLines)
   }
 
-  private _fireContentChanges(): void {
-    let { textDocument } = this
+  private _fireContentChanges(): boolean {
     let { cursor } = events
+    let { textDocument } = this
     try {
-      let content = this.getDocumentContent()
       let endOffset = null
+      // consider cursor position.
       if (cursor && cursor.bufnr == this.bufnr) {
         endOffset = this.getEndOffset(cursor.lnum, cursor.col, cursor.insert)
       }
+      let content = this.getDocumentContent()
       let change = getChange(textDocument.getText(), content, endOffset)
       if (change == null) return
-      this.createDocument()
-      let { version, uri } = this
       let start = textDocument.positionAt(change.start)
       let end = textDocument.positionAt(change.end)
       let original = textDocument.getText(Range.create(start, end))
+      this._version = this._version + 1
+      this.syncLines = this.lines
       let changes = [{
         range: { start, end },
         rangeLength: change.end - change.start,
@@ -262,42 +269,15 @@ export default class Document {
       this._onDocumentChange.fire({
         bufnr: this.bufnr,
         original,
-        textDocument: { version, uri },
+        textDocument: { version: this.version, uri: this.uri },
         contentChanges: changes
       })
-      this._words = this.chars.matchKeywords(this.textDocument.getText())
+      this._words = this.chars.matchKeywords(content)
+      return true
     } catch (e) {
       logger.error(e.message)
     }
-  }
-
-  /**
-   * Buffer number
-   */
-  public get bufnr(): number {
-    return this.buffer.id
-  }
-
-  /**
-   * Content of textDocument.
-   */
-  public get content(): string {
-    return this.textDocument.getText()
-  }
-
-  /**
-   * Coverted filetype.
-   */
-  public get filetype(): string {
-    return this._filetype
-  }
-
-  public get uri(): string {
-    return this._uri
-  }
-
-  public get version(): number {
-    return this.textDocument ? this.textDocument.version : null
+    return false
   }
 
   public async applyEdits(edits: TextEdit[]): Promise<void> {
@@ -308,47 +288,76 @@ export default class Document {
     edits.forEach(edit => {
       edit.newText = edit.newText.replace(/\r/g, '')
     })
-    let current = this.lines.join('\n') + (this.eol ? '\n' : '')
+    let current = this.getDocumentContent()
     let textDocument = TextDocument.create(this.uri, this.filetype, 1, current)
     // apply edits to current textDocument
     let applied = TextDocument.applyEdits(textDocument, edits)
     // could be equal sometimes
     if (current !== applied) {
-      let newLines = applied.split('\n')
-      if (this.eol && newLines[newLines.length - 1] == '') {
-        newLines = newLines.slice(0, -1)
-      }
+      let newLines = (this.eol && applied.endsWith('\n') ? applied.slice(0, -1) : applied).split('\n')
       let d = diffLines(this.lines, newLines)
-      await this.buffer.setLines(d.replacement, {
-        start: d.start,
-        end: d.end,
-        strictIndexing: false
-      })
-      // can't wait vim sync buffer
-      this.lines = newLines
-      this.forceSync()
+      let release = await this.mutex.acquire()
+      try {
+        let res = await this.nvim.call('coc#util#set_lines', [this.bufnr, d.replacement, d.start, d.end])
+        this._changedtick = res.changedtick
+        // can't wait vim sync buffer
+        this.lines = newLines
+        // res.lines
+        this.fireContentChanges.clear()
+        this._fireContentChanges()
+        release()
+        // could be user type during applyEdits.
+        if (!equals(newLines, res.lines)) {
+          process.nextTick(() => {
+            this.lines = res.lines
+            this.fireContentChanges.clear()
+            this._fireContentChanges()
+          })
+        }
+      } catch (e) {
+        logger.error('Error on applyEdits: ', e)
+        release()
+      }
     }
   }
 
-  public changeLines(lines: [number, string][], sync = true): void {
-    let { nvim } = this
+  public async changeLines(lines: [number, string][]): Promise<void> {
     let filtered: [number, string][] = []
+    let newLines = this.lines.slice()
     for (let [lnum, text] of lines) {
-      if (this.lines[lnum] != text) {
+      if (newLines[lnum] != text) {
         filtered.push([lnum, text])
-        this.lines[lnum] = text
+        newLines[lnum] = text
       }
     }
     if (!filtered.length) return
-    nvim.call('coc#util#change_lines', [this.bufnr, filtered], true)
-    this.ignoreChange = true
-    if (sync) this.forceSync()
+    let release = await this.mutex.acquire()
+    try {
+      let res = await this.nvim.call('coc#util#change_lines', [this.bufnr, filtered])
+      if (res != null) {
+        this.lines = newLines
+        this._changedtick = res.changedtick
+        this.fireContentChanges.clear()
+        this._fireContentChanges()
+        if (!equals(newLines, res.lines)) {
+          process.nextTick(() => {
+            this.lines = res.lines
+            this.fireContentChanges.clear()
+            this._fireContentChanges()
+          })
+        }
+      }
+      release()
+    } catch (e) {
+      release()
+    }
   }
 
   /**
    * Force document synchronize and emit change event when necessary.
    */
   public forceSync(): void {
+    if (this.mutex.busy) return
     this.fireContentChanges.clear()
     this._fireContentChanges()
   }
@@ -426,41 +435,29 @@ export default class Document {
     return Range.create(position.line, start, position.line, end)
   }
 
-  private gitCheck(): void {
-    let { uri } = this
-    if (!uri.startsWith('file') || this.buftype != '') return
-    let filepath = URI.parse(uri).fsPath
-    isGitIgnored(filepath).then(isIgnored => {
-      this.isIgnored = isIgnored
-    }, () => {
-      this.isIgnored = false
-    })
+  /**
+   * Synchronized textDocument.
+   */
+  public get textDocument(): TextDocument {
+    let { version, filetype, uri } = this
+    return TextDocument.create(uri, filetype, version, this.content)
   }
 
-  private createDocument(changeCount = 1): void {
-    let { version, uri, filetype } = this
-    version = version + changeCount
-    this.textDocument = TextDocument.create(
-      uri,
-      filetype,
-      version,
-      this.getDocumentContent()
-    )
-  }
-
+  /**
+   * Used by vim for fetch new lines.
+   */
   private async _fetchContent(): Promise<void> {
     if (!this.env.isVim || !this._attached) return
-    let { nvim, buffer } = this
-    let { id } = buffer
-    let o = (await nvim.call('coc#util#get_content', id))
-    if (!o) return
-    let { content, changedtick } = o
-    if (this._changedtick == changedtick) return
-    this._changedtick = changedtick
-    let newLines: string[] = content.split('\n')
-    this.lines = newLines
-    this.fireContentChanges.clear()
-    this._fireContentChanges()
+    let { nvim, bufnr, changedtick } = this
+    let release = await this.mutex.acquire()
+    let o = await nvim.call('coc#util#get_buf_lines', [bufnr, changedtick])
+    if (o && o.changedtick >= this._changedtick) {
+      this._changedtick = o.changedtick
+      this.lines = o.lines
+      this.fireContentChanges.clear()
+      this._fireContentChanges()
+    }
+    release()
   }
 
   /**
@@ -471,11 +468,13 @@ export default class Document {
     if (this.env.isVim) {
       if (currentLine) {
         let change = await this.nvim.call('coc#util#get_changeinfo', []) as ChangeInfo
-        if (change.changedtick == this._changedtick) return
-        let { lines } = this
+        if (change.changedtick < this._changedtick) return
         let { lnum, line, changedtick } = change
+        let newLines = this.lines.slice()
         this._changedtick = changedtick
-        lines[lnum - 1] = line
+        if (newLines[lnum - 1] == line) return
+        newLines[lnum - 1] = line
+        this.lines = newLines
         this.forceSync()
       } else {
         this.fetchContent.clear()
@@ -492,8 +491,8 @@ export default class Document {
    */
   public getSymbolRanges(word: string): Range[] {
     this.forceSync()
-    let { textDocument } = this
     let res: Range[] = []
+    let { textDocument } = this
     let content = textDocument.getText()
     let str = ''
     for (let i = 0, l = content.length; i < l; i++) {
@@ -537,23 +536,11 @@ export default class Document {
   }
 
   /**
-   * Get cwd of this document.
-   *
-   * @deprecated won't work when buffer not in current tab.
-   */
-  public async getcwd(): Promise<string> {
-    let wid = await this.nvim.call('bufwinid', this.buffer.id)
-    if (wid == -1) return await this.nvim.call('getcwd')
-    return await this.nvim.call('getcwd', wid)
-  }
-
-  /**
    * Real current line
    */
   public getline(line: number, current = true): string {
     if (current) return this.lines[line] || ''
-    let lines = this.textDocument.getText().split(/\r?\n/)
-    return lines[line] || ''
+    return this.syncLines[line] || ''
   }
 
   /**
@@ -591,7 +578,7 @@ export default class Document {
 
   /**
    * Get end offset from cursor position.
-   * For normal mode, use offset -1 when possible
+   * For normal mode, use offset - 1 when possible
    */
   public getEndOffset(lnum: number, col: number, insert: boolean): number {
     let total = 0
@@ -622,11 +609,8 @@ export default class Document {
    * @internal
    */
   public setFiletype(filetype: string): void {
-    let { uri, version } = this
     this._filetype = this.convertFiletype(filetype)
-    version = version ? version + 1 : 1
-    let textDocument = TextDocument.create(uri, this.filetype, version, this.content)
-    this.textDocument = textDocument
+    this._version = this._version + 1
   }
 
   /**
@@ -643,7 +627,12 @@ export default class Document {
       }
     }
     let lines = this.lines.length > 30000 ? this.lines.slice(0, 30000) : this.lines
+    // TODO not parse words
     this._words = this.chars.matchKeywords(lines.join('\n'))
+  }
+
+  public get attached(): boolean {
+    return this._attached
   }
 
   /**
@@ -659,10 +648,6 @@ export default class Document {
     this.fireContentChanges.clear()
     this._onDocumentChange.dispose()
     this._onDocumentDetach.dispose()
-  }
-
-  public get attached(): boolean {
-    return this._attached
   }
 
   /**
