@@ -1,7 +1,8 @@
 import { Neovim } from '@chemzqm/neovim'
 import Notification from './notification'
 import { NotificationPreferences, Progress } from '../types'
-import { CancellationToken, CancellationTokenSource, Disposable } from 'vscode-languageserver-protocol'
+import { CancellationToken, CancellationTokenSource } from 'vscode-languageserver-protocol'
+import events from '../events'
 
 export interface ProgressOptions<R> {
   title?: string
@@ -11,26 +12,34 @@ export interface ProgressOptions<R> {
 
 export default class ProgressNotification<R> extends Notification {
   private tokenSource: CancellationTokenSource
-  protected disposables: Disposable[] = []
   constructor(nvim: Neovim, private option: ProgressOptions<R>) {
     super(nvim, {
       content: '\n',
       close: option.cancellable == true,
       title: option.title
-    })
+    }, false)
+    events.on('BufWinLeave', bufnr => {
+      if (bufnr == this.bufnr) {
+        if (this.tokenSource) {
+          this.tokenSource.cancel()
+        }
+        this.dispose()
+      }
+    }, null, this.disposables)
   }
 
   public async show(preferences: Partial<NotificationPreferences>): Promise<R> {
-    let shown = await super.show(Object.assign({ minWidth: preferences.minProgressWidth || 30, progress: 1 }, preferences))
-    if (!shown) return undefined
     let { task } = this.option
     let tokenSource = this.tokenSource = new CancellationTokenSource()
+    this.disposables.push(tokenSource)
     let total = 0
     let res = await new Promise<R>((resolve, reject) => {
       tokenSource.token.onCancellationRequested(() => {
-        this.dispose()
         resolve(undefined)
       })
+      super.show(Object.assign({ minWidth: preferences.minProgressWidth || 30, progress: 1 }, preferences)).then(shown => {
+        if (!shown) reject(new Error('Failed to create float window'))
+      }).catch(reject)
       task({
         report: p => {
           if (!this.bufnr) return
@@ -43,23 +52,26 @@ export default class ProgressNotification<R> extends Notification {
           this.nvim.call('setbufline', [this.bufnr, 2, text], true)
         }
       }, tokenSource.token).then(res => {
-        this.dispose()
+        if (this._disposed) return
+        setTimeout(() => {
+          this.dispose()
+        }, 100)
         resolve(res)
       }, err => {
+        if (this._disposed) return
         this.dispose()
-        reject(err)
+        if (err instanceof Error) {
+          reject(err)
+        } else {
+          resolve(undefined)
+        }
       })
     })
     return res
   }
 
   public dispose(): void {
-    let { tokenSource } = this
-    if (tokenSource) {
-      this.tokenSource = undefined
-      tokenSource.cancel()
-      tokenSource.dispose()
-    }
     super.dispose()
+    this.tokenSource = undefined
   }
 }
