@@ -133,7 +133,7 @@ export default class Plugin extends EventEmitter {
       let [fn, repeat] = keymap
       let res = await Promise.resolve(fn())
       if (repeat) await nvim.command(`silent! call repeat#set("\\<Plug>(coc-${key})", -1)`)
-      return res || defaultReturn
+      return res ?? defaultReturn
     })
     this.addAction('registExtensions', async (...folders: string[]) => {
       for (let folder of folders) {
@@ -208,6 +208,9 @@ export default class Plugin extends EventEmitter {
     })
     this.addAction('startCompletion', async option => {
       await completion.startCompletion(option)
+    })
+    this.addAction('stopCompletion', () => {
+      completion.stop(false)
     })
     this.addAction('sourceStat', () => {
       return sources.sourceStats()
@@ -464,6 +467,7 @@ export default class Plugin extends EventEmitter {
     extensions.dispose()
     listManager.dispose()
     workspace.dispose()
+    window.dispose()
     sources.dispose()
     services.stopAll()
     services.dispose()
