@@ -1,4 +1,4 @@
-import { SelectionRange, CancellationToken, CodeAction, CodeActionContext, CodeActionKind, CodeLens, Color, ColorInformation, ColorPresentation, Command, CompletionContext, CompletionItem, CompletionList, Definition, DocumentHighlight, DocumentLink, DocumentSymbol, FoldingRange, FormattingOptions, Hover, Location, Position, Range, SignatureHelp, SymbolInformation, TextEdit, WorkspaceEdit, Event, DefinitionLink } from 'vscode-languageserver-protocol'
+import { CancellationToken, CodeAction, CodeActionContext, CodeActionKind, CodeLens, Color, ColorInformation, ColorPresentation, Command, CompletionContext, CompletionItem, CompletionList, Definition, DefinitionLink, DocumentHighlight, DocumentLink, DocumentSymbol, Event, FoldingRange, FormattingOptions, Hover, Location, Position, Range, SelectionRange, SignatureHelp, SignatureHelpContext, SymbolInformation, TextEdit, WorkspaceEdit } from 'vscode-languageserver-protocol'
 import { TextDocument } from 'vscode-languageserver-textdocument'
 import { URI } from 'vscode-uri'
 
@@ -129,7 +129,7 @@ export interface DefinitionProvider {
     document: TextDocument,
     position: Position,
     token: CancellationToken
-  ): ProviderResult<Definition>
+  ): ProviderResult<Definition | DefinitionLink[]>
 }
 
 /**
@@ -161,7 +161,8 @@ export interface SignatureHelpProvider {
   provideSignatureHelp(
     document: TextDocument,
     position: Position,
-    token: CancellationToken
+    token: CancellationToken,
+    context: SignatureHelpContext
   ): ProviderResult<SignatureHelp>
 }
 
@@ -183,7 +184,7 @@ export interface TypeDefinitionProvider {
     document: TextDocument,
     position: Position,
     token: CancellationToken
-  ): ProviderResult<Definition>
+  ): ProviderResult<Definition | DefinitionLink[]>
 }
 
 /**
@@ -282,7 +283,7 @@ export interface ImplementationProvider {
     document: TextDocument,
     position: Position,
     token: CancellationToken
-  ): ProviderResult<Definition>
+  ): ProviderResult<Definition | DefinitionLink[]>
 }
 
 /**
@@ -423,7 +424,7 @@ export interface DocumentRangeFormattingEditProvider {
  *
  * A code action can be any command that is [known](#commands.getCommands) to the system.
  */
-export interface CodeActionProvider {
+export interface CodeActionProvider<T extends CodeAction = CodeAction> {
   /**
    * Provide commands for the given document and range.
    *
@@ -441,6 +442,18 @@ export interface CodeActionProvider {
     context: CodeActionContext,
     token: CancellationToken
   ): ProviderResult<(Command | CodeAction)[]>
+
+  /**
+   * Given a code action fill in its [`edit`](#CodeAction.edit)-property. Changes to
+   * all other properties, like title, are ignored. A code action that has an edit
+   * will not be resolved.
+   *
+   * @param codeAction A code action.
+   * @param token A cancellation token.
+   * @return The resolved code action or a thenable that resolves to such. It is OK to return the given
+   * `item`. When no result is returned, the given `item` will be used.
+   */
+  resolveCodeAction?(codeAction: T, token: CancellationToken): ProviderResult<T>
 }
 
 /**
